Update metodo. Se mueve la comprobacion de autorizacion a la capa controladora

+++ jee_ecp/src/main/java/es/upm/miw/jee/ecp/controllers/ejb/RemoveTemaControllerEjb.java
@@ -5,7 +5,12 @@ import es.upm.miw.jee.ecp.models.daos.jpa.DaoJpaFactory;
 
 public class RemoveTemaControllerEjb implements RemoveTemaController {
 
-	public void removeTema(Integer temaId) {
-		DaoJpaFactory.getFactory().getTemaDao().deleteById(temaId);
+	public Boolean removeTema(Integer temaId, String autorizacionCode) {
+		if (autorizacionCode.equalsIgnoreCase(AUTH_CODE)){
+			DaoJpaFactory.getFactory().getTemaDao().deleteById(temaId);
+			return true;
+		}else{
+			return false;
+		}
 	}
 }
